fix: rename percentage diff helper
