+++ src/lib/helpers/get-product-price.ts
@@ -1,5 +1,5 @@
 import { HttpTypes } from "@medusajs/types"
-import { getPercentageDiff } from "./get-precentage-diff"
+import { getPercentageDiff } from "./get-percentage-diff"
 import { convertToLocale } from "./money"
 import { BaseHit, Hit } from "instantsearch.js"
